Hello friends,
This is my first guided project, and I have done additional research apart from the lessons taught about libraries, functions, and methods to bring you a quality project. However, I encountered limitations such as question 4, for which I could not find a solution, and question 5, which (I believe) could not be completed properly due to limitations in data processing capacity or errors in the data. I would appreciate your feedback on the answered questions and your support on the question I could not answer. I apologize for the language used in the project, but I also want the participation of the Spanish-speaking community.

Thank you very much for reading,
Francisco.